feat: 使用者離開操作頁面後刪除所有檔案
調整項目：
1. 一旦使用者回到主頁，代表操作已完成，刪除所有檔案

隱患：
1. issue #108

issue #98
