+++ Template/general/function_home.html
@@ -62,12 +62,12 @@
                     contentType: false,
                     processData: false,
                     success: function (finlish) {
-                            if(finlish == true){
-                                window.location = '/{% block custom_url %}json_parser{% endblock custom_url %}/custom/csv_name='+csv_name+'/';
-                            } else{
-                                alert('檔案上傳失敗，請稍後再試');
-                                window.location = "{% url 'home' %}";
-                            }
+                        if(finlish == true){
+                            window.location = '/{% block custom_url %}json_parser{% endblock custom_url %}/custom/csv_name='+csv_name+'/';
+                        } else{
+                            alert('檔案上傳失敗，請稍後再試');
+                            window.location = "{% url 'home' %}";
+                        }
                     },
                     error: function (xhr, ajaxOptions, thrownError) {
                         var jsonObj = JSON.parse(xhr.responseText);
