+++ home/urls.py
@@ -3,4 +3,5 @@ from . import views
 
 urlpatterns = [
     path('', views.index, name='home'),
+    path('initialize/', views.InitializeView.as_view()),
 ]
