+++ k_Anonymity/urls.py
@@ -12,7 +12,6 @@ urlpatterns = [
     path('File_List_<str:method>/', general_views.FileListView.as_view()),
     path('Download_Output/csv_name=<str:csv_name>/', general_views.DownloadView.as_view()),
     path('CSV_View_<str:method>/', general_views.DisplayCsvView.as_view()),
-    path('delete_file_<str:method>/', general_views.DeleteFileView.as_view()),
     path('custom/csv_name=<str:csv_name>/', json_parser_views.CustomView.as_view()),
     path('finish/csv_name=<str:csv_name>/', general_views.FinishView.as_view()),
     path('utility_page/csv_name=<str:csv_name>/', general_views.UtilityPageView.as_view()),
